+++ app/src/main/java/com/example/tateknew/data/MeterReading.kt
@@ -2,13 +2,14 @@ package com.example.tateknew.data
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 import java.util.Date
 
 @Entity(tableName = "meter_readings")
 data class MeterReading(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val mtrId: Long,
-    val currentReading: Double,
+    val currentReading: BigDecimal,
     val photoPath: String,
     val latitude: Double,
     val longitude: Double,
